requirements.txt
annotated-types==0.7.0
anyio==4.12.0
arabic-reshaper==3.0.0
asgiref==3.11.0
asn1crypto==1.5.1
cachetools==6.2.4
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
colorama==0.4.6
cryptography==46.0.3
cssselect2==0.8.0
distro==1.9.0
Django==6.0
freetype-py==2.5.1
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
lxml==6.0.2
oscrypto==1.3.0
pillow==12.0.0
proto-plus==1.27.0
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycairo==1.29.0
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyHanko==0.32.0
pyhanko-certvalidator==0.29.0
pyparsing==3.3.1
pypdf==6.5.0
python-bidi==0.6.7
python-dotenv==1.2.1
PyYAML==6.0.3
reportlab==4.4.7
requests==2.32.5
rlPyCairo==0.4.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.5
svglib==1.6.0
tenacity==9.1.2
tinycss2==1.5.1
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
tzlocal==5.3.1
uritemplate==4.2.0
uritools==6.0.1
urllib3==2.6.2
webencodings==0.5.1
websockets==15.0.1
xhtml2pdf==0.2.17
